pyproject.toml
[tool.poetry]
name = "zoneit"
version = "0.1.9"
description = ""
authors = ["Hugo Caron <hca443@gmail.com>"]
readme = "README.md"
packages = [{ include = "zoneit" }]

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
pydantic = "^2.9.0"
fastapi = { extras = ["standard"], version = "^0.113.0" }
httpx = "^0.27.2"
redis = "^5.0.8"
asyncio = "^3.4.3"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.3"
ruff-lsp = "^0.0.55"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]

[tool.ruff]
line-length = 88
indent-width = 4
lint.extend-select = ["I"]
